Final syntax fixes for phase lag and efficacy
interfaces fixed to match changes, efficacy and phase lag complete pending review

+++ Battelle.EPA.WideAreaDecon.Model/IncidentCommand/ILaborCostCalculator.cs
@@ -3,8 +3,8 @@
     public interface ILaborCostCalculator
     {
         public double CalculateOnSiteDays(double _numberTeams, double surfaceAreaToBeSourceReduced, double personnelRoundTripDays,
-            double _surfaceAreaToBeHepa, double _surfaceAreaToBeWiped, double numberLabs, double sampleTimeTransmitted);
+            double _surfaceAreaToBeHepa, double _surfaceAreaToBeWiped, int numberLabs, double sampleTimeTransmitted);
         public double CalculateLaborCost(double _numberTeams, double surfaceAreaToBeSourceReduced, double personnelRoundTripDays,
-            double _surfaceAreaToBeHepa, double _surfaceAreaToBeWiped, double numberLabs, double sampleTimeTransmitted);
+            double _surfaceAreaToBeHepa, double _surfaceAreaToBeWiped, int numberLabs, double sampleTimeTransmitted);
     }
 }
