+++ Battelle.EPA.WideAreaDecon.Model/IncidentCommand/SuppliesCostCalculator.cs
@@ -18,7 +18,7 @@
         }
 
         public double CalculateSuppliesCost(double _numberTeams, double surfaceAreaToBeSourceReduced, double personnelRoundTripDays, double _surfaceAreaToBeHepa
-            ,double _surfaceAreaToBeWiped, double numberLabs, double sampleTimeTransmitted)
+            ,double _surfaceAreaToBeWiped, int numberLabs, double sampleTimeTransmitted)
         {
             var onSiteDays = _laborCostCalculator.CalculateOnSiteDays(   _numberTeams,  surfaceAreaToBeSourceReduced,  personnelRoundTripDays,
              _surfaceAreaToBeHepa, _surfaceAreaToBeWiped,  numberLabs,  sampleTimeTransmitted);
